+++ src/components/action-popover/action-popover-menu-button/action-popover-menu-button.component.tsx
@@ -11,7 +11,7 @@ import { IconType } from "../../icon";
 
 export type ActionPopoverMenuButtonAria = {
   "aria-haspopup": string;
-  "aria-label": string;
+  "aria-label"?: string;
   "aria-labelledby"?: string;
   "aria-describedby"?: string;
   "aria-controls": string;