+++ src/components/button/button.test.tsx
@@ -3,18 +3,6 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import Button from "./button.component";
 
-import Logger from "../../__internal__/utils/logger";
-
-let loggerSpy: jest.SpyInstance;
-
-beforeEach(() => {
-  loggerSpy = jest.spyOn(Logger, "deprecate").mockImplementation(() => {});
-});
-
-afterEach(() => {
-  loggerSpy.mockRestore();
-});
-
 test("renders with text children", () => {
   render(<Button>foo</Button>);
 
@@ -303,18 +291,6 @@ test("sets the text content of the 'subtext' element when the 'size' prop is 'la
   expect(subtext).toBeVisible();
 });
 
-test("sets the 'className' attribute correctly when a custom value is passed to the 'className' prop", () => {
-  render(<Button className="foo">bar</Button>);
-
-  const button = screen.getByRole("button", { name: "bar" });
-
-  expect(button).toHaveClass("foo");
-  expect(loggerSpy).toHaveBeenCalledWith(
-    "The 'className' prop has been deprecated and will soon be removed from the 'Button' component.",
-  );
-  expect(loggerSpy).toHaveBeenCalledTimes(1);
-});
-
 test("calls onClick when an 'href' is passed and the space key is pressed", async () => {
   const user = userEvent.setup();
   const clickMock = jest.fn();
@@ -379,7 +355,7 @@ test("accepts ref as a ref callback", () => {
 
 test("sets ref to empty after unmount", () => {
   const mockRef = { current: null };
-  const { unmount } = render(<Button className="foo">bar</Button>);
+  const { unmount } = render(<Button>bar</Button>);
 
   unmount();