+++ src/components/tabs/tabs.test.tsx
@@ -94,24 +94,6 @@ test("should throw an error if rendered with `headerWidth` prop and `position` p
   consoleSpy.mockRestore();
 });
 
-test("passes the `className` prop down to the element", () => {
-  render(
-    <Tabs className="custom-class-1 custom-class-2">
-      <Tab tabId="foo" />
-    </Tabs>,
-  );
-
-  expect(screen.getByTestId("tabs")).toHaveClass(
-    "custom-class-1",
-    "custom-class-2",
-  );
-
-  expect(loggerSpy).toHaveBeenCalledWith(
-    "The 'className' prop has been deprecated and will soon be removed from the 'Tabs' component.",
-  );
-  expect(loggerSpy).toHaveBeenCalledTimes(1);
-});
-
 test("the `selectedTabId` prop determines which child `Tab` is displayed", () => {
   render(
     <Tabs selectedTabId="tab-2">